moving auto markers
